Backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.27
psycopg2>=2.9.9
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
alembic>=1.13.1 
bcrypt==4.0.1
email-validator==2.2.0